Removed type and density members from example
The usage of type and icon was incorrect. Those members should used to "discriminate" within a set of icons. The same with `type`: the semantics are "only use this icon if you support this type". The use case is for WebP or future file formats that are not universally supported.